Reorg ascent module and commands
Reorg ascent module by adding an AscentLog class. Adding this
class allows for significant simplification of the log_ascent and
drop_ascent commands and brings them more in line with each other.
It also allows for the removal of the separate .known_crags file.

Note that despite this reorg, the actual experience of using the
commands was preserved (or at least it was intended to be).